Fixed more warnings (and one new error) ;)
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@2459 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ DarkMod/AI/Tasks/PathAnimTask.cpp
@@ -54,7 +54,7 @@ void PathAnimTask::Init(idAI* owner, Subsystem& subsystem)
 		subsystem.FinishTask();
 	}
 
-	float blendIn = path->spawnArgs.GetInt("blend_in");
+	int blendIn = path->spawnArgs.GetInt("blend_in");
 	
 	// Synchronise the leg channel
 	owner->Event_OverrideAnim(ANIMCHANNEL_LEGS);
